Add switchable U256/U512 backend via compile-time feature flags
Introduce a #[repr(transparent)] newtype wrapper for U256 and U512 in
ethrex-common that delegates to a backend selected by feature flag.
This enables benchmarking different 256-bit integer implementations
for zkVM proving performance without codebase-wide refactoring.

Two backends are supported:
- uint-ethereum-types (default): ethereum_types::U256
- uint-ruint: ruint::Uint<256, 4>

To switch backends, change the feature in the workspace Cargo.toml.

The wrapper provides ~50 methods and ~20 trait implementations covering
arithmetic, bitwise operations, byte conversion, serialization (RLP,
serde), cross-type arithmetic, and BigEndianHash interop. All methods
are #[inline] for zero-cost abstraction.

Migration changes across ~40 files replace direct ethereum_types::U256
imports with ethrex_common::U256 and update internal .0 field access
to use from_limbs()/as_limbs() methods.

+++ crates/vm/levm/src/opcode_handlers/bitwise_comparison.rs
@@ -233,7 +233,7 @@ impl OpcodeHandler for OpShlHandler {
             .stack
             .push(match u8::try_from(shift_amount) {
                 #[expect(clippy::arithmetic_side_effects, reason = "U256 shift by u8 is safe")]
-                Ok(shift_amount) => value << shift_amount,
+                Ok(shift_amount) => value << usize::from(shift_amount),
                 Err(_) => U256::zero(),
             })?;
 
@@ -253,7 +253,7 @@ impl OpcodeHandler for OpShrHandler {
             .stack
             .push(match u8::try_from(shift_amount) {
                 #[expect(clippy::arithmetic_side_effects, reason = "U256 shift by u8 is safe")]
-                Ok(shift_amount) => value >> shift_amount,
+                Ok(shift_amount) => value >> usize::from(shift_amount),
                 Err(_) => U256::zero(),
             })?;
 
@@ -273,8 +273,8 @@ impl OpcodeHandler for OpSarHandler {
         vm.current_call_frame
             .stack
             .push(match (u8::try_from(shift_amount), value.bit(255)) {
-                (Ok(shift_amount), false) => value >> shift_amount,
-                (Ok(shift_amount), true) => !(!value >> shift_amount),
+                (Ok(shift_amount), false) => value >> usize::from(shift_amount),
+                (Ok(shift_amount), true) => !(!value >> usize::from(shift_amount)),
                 (Err(_), false) => U256::zero(),
                 (Err(_), true) => U256::MAX,
             })?;
